feat(sdk): `event_cache::Pagination::run_backwards` paginates in the event cache.
This patch updates `Pagination::run_backwards_impl` to paginate in the
event cache. The flow is now as follow:

- backwards pagination tries to load and to insert a new previous chunk
   from the store
  - if the new chunk contains events, they are returned, pagination done
  - if the new chunk is a gap, the flow continues
- (as previously) check for a prev batch token (it exists in the newly
  inserted gap)
- (as previously) run a network request, replace the gap by the new
  events
- etc.

The new part is to load and to insert a new previous chunk. The rest
is stays the same. The code has been moved in code to keep the lock
releases happy and to clarify the code.